Added controllers, services and repositories for some classes

+++ src/main/java/com/recipes/api/dtos/DietDto.java
@@ -14,7 +14,7 @@ import jakarta.validation.constraints.Size;
 @AllArgsConstructor
 @NoArgsConstructor
 public class DietDto {
-    private Long id;
+    private Integer id;
 
     @NotBlank(message = "Diet name is mandatory")
     @Size(max = 100, message = "The maximum length is 100 characters")
